==> 02_CORDA_integration.ipynb <==
import pandas as pd
import cobra

from cobra.io import load_model

import numpy as np
import pickle

import os.path

Ponovno preberemo model Recon3D, H-GEM:

model = cobra.io.read_sbml_model(os.path.join('models','Recon3D.xml'))

model_H_gem = cobra.io.read_sbml_model(os.path.join('models','Human-GEM.xml'))

print(model.genes[:5])
print(model_H_gem.genes[:5])

# Observe methods for model_H_gem.genes
print(dir(model_H_gem.genes[0]))

print(model_H_gem.genes[0].id)

H-gem uporablja ENS genske oznake, medtem RECON3D genske simbole.

Pripravimo si slovarja `id gena` --> `ime gena` in obratno:

gene_id_name = {} # seznam ID genov
gene_name_id = {} # seznam imen genov

for g in model.genes[1:]:
    gene_id_name[g.id] = g.name
    gene_name_id[g.name] = g.id

list(gene_id_name.items())[:5] # prvih 5 elementov v slovarju, ki ID gena poveže z imenom

list(gene_name_id.items())[:5] # prvih 5 elementov v slovarju, ki ime gena poveže z ID

Preberimo podatke, ki smo jih prej shranili:

df_control = pd.read_csv(os.path.join('data','data_normal.txt'))
df_tumor = pd.read_csv(os.path.join('data','data_tumor.txt'))

Imenom genov pripišimo še ustrezne ID-je:

df_control['gene_id'] = df_control['gene'].map(lambda x: gene_name_id[x] if x in gene_name_id else "")
df_tumor['gene_id'] = df_tumor['gene'].map(lambda x: gene_name_id[x] if x in gene_name_id else "")

df_control

Ohranimo samo tiste vnose, ki jih imamo v modelu Recon3D:

df_control
print("Genes successfully mapped: ", len(df_control[df_control['gene_id'] != '']) / len(df_control))

df_control_sub_recon = df_control[df_control['gene_id'] != ""]
df_tumor_sub_recon = df_tumor[df_tumor['gene_id'] != ""]

len(df_control_sub_recon), len(df_tumor_sub_recon)

df_control_sub_recon.describe()

df_tumor_sub_recon.describe()

In v modelu H-GEM

# Create a list of all genes in H-GEM
model_H_gem_genes = [g.id for g in model_H_gem.genes]
model_H_gem_genes

# Subsett all df_control rows where col 'ENS_ID' is not present in model_H_gem genes id
index = df_control['ENS_ID'].isin(model_H_gem_genes)

# Get the percentage of genes that are present in H-GEM
print("Genes successfully mapped: ", len(df_control[index]) / len(df_control))


df_control_sub_H_gem = df_control[index]
df_tumor_sub_H_gem = df_tumor[index]
df_tumor_sub_H_gem

len(df_control_sub_H_gem), len(df_tumor_sub_H_gem)

df_control_sub_H_gem.describe()

df_tumor_sub_H_gem.describe()

Who captures more DE genes?

print('Genes in RECON3D: ', len(df_tumor_sub_recon))
print('Genes in H-GEM: ', len(df_tumor_sub_H_gem))

### Integracija transkriptomskih podatkov v model

Uporabili bomo algoritem CORDA:

* [https://resendislab.github.io/corda/](https://resendislab.github.io/corda/)
* [https://github.com/resendislab/corda](https://github.com/resendislab/corda)

Funkcija za rekonstrukcijo kot argument sprejme model in preslikavo `confidence`, ki za vsako reakcijo poda mero zaupanja, da je vsebovana v modelu:
```
Allowed confidence values are -1 (absent/do not include), 0 (unknown), 1 (low confidence), 2 (medium confidence) and 3 (high confidence).
```

`confidence` si lahko pripravimo z uporabo CORDA funkcije `reaction_confidence`, ki preslika stopnje izražanja genov (po enaki lestvici kot zgoraj) v stopnje aktivnosti reakcij z uporabo pravil GPR.


Dodatno lahko nastavimo zmožnost modela, da producira biomaso:
```
met_prod (Optional[list]): Additional metabolic targets that have to be
            achieved by the model. Can be a single object or list of objects.
            List elements can be given in various forms:
            (1) A string naming a metabolite in the model will ensure that the
            metabolite can be produced.
            For instance: met_prod = ["pi_c", "atp_c"]
            (2) A dictionary of metabolite -> int will define an irreversible
            reaction that must be able to carry flux.
            For instance: met_prod = {"adp_c": -1, "pi_c": -1, "atp_c": 1}
            (3) A string representation of a reversible or irreversible
            reaction that must be able to carry flux.
            For instance: met_prod = "adp_c + pi_c -> atp_c"

```
Primer priprave niza `met_prod`:
```Python
met_prod = model.reactions.BIOMASS_maintenance.build_reaction_string()
met_prod = met_prod.replace("-->", "->")
```


biomass = model.reactions.BIOMASS_maintenance.build_reaction_string().replace("-->", "->")
biomass

### Preslikava izražanja genov v prisotnosti reakcij

from corda import reaction_confidence
from corda import CORDA

Pomagali si bomo s funkcijo, ki bo preslikala izražanje genov v vrednosti -1, 0, 1, 2 in 3 preko vnaprej definiranih pragov.

def val_to_score(val, thr0, thr1, thr2, thr3):
    if val < thr0:
        return -1
    elif val < thr1:
        return 0
    elif val < thr2:
        return 1
    elif val < thr3:
        return 2
    else:
        return 3

#### Parametri rekonstrukcije
Pragove bomo definirali preko odstotne aktivnosti genov v vzrocu. Obstaja veliko različnih strategij, kako se tega problema lotiti. Nekaj več o tem si lahko preberete v [članku](https://doi.org/10.1371/journal.pcbi.1007185).

require_biomass = True
perc0 = 25 # below this ... inactive
perc1 = 50 # else, below this ... not sure
perc2 = 75 # else, below this ...  low confidence
perc3 = 95 # else, below this ... medium confidence
# else high confidence (above perc3)

folder = os.path.join('models','biomass') if require_biomass else os.path.join('models','no_biomass')
folder

### Rangiranje
Za vsako skupino bomo določili lastne pragove glede na aktivnosti genov:

# Deffine function for confidence score
def calculate_confidence(df, perc0, perc1, perc2, perc3, model, genes='gene_id'):
    
    thr0 = np.percentile(df.value, perc0)
    thr1 = np.percentile(df.value, perc1)
    thr2 = np.percentile(df.value, perc2)
    thr3 = np.percentile(df.value, perc3)
    
    # Najprej preslikajmo intenziteto izražanja genov v skupine
    gene_conf = {g_id: val_to_score(val, thr0, thr1, thr2, thr3) for (g_id, val) in zip(df[genes], df['value'])}

    # Potem pa skupine genov uporabimo pri razvrščanju reakcij v skupine z uporabo
    # CORDA funkcije `reaction_confidence`, ki pri preslikavi upošteva pravila GPR.
    conf = {}
    for r in model.reactions:
        conf[r.id] = reaction_confidence(r, gene_conf)

    return conf

conf_control = calculate_confidence(
    df = df_control_sub_recon,
    perc0 = perc0,
    perc1 = perc1,
    perc2 = perc2,
    perc3 = perc3,
    model = model
    )
conf_control

Potem poženimo integracijo z algoritmom:

**POZOR: časovno potratno! Spodnje celice lahko preskočiš in podatke prebereš iz datoteke (spodaj).**

if require_biomass:
    opt_control = CORDA(model, conf_control, met_prod = biomass)
else:
    opt_control = CORDA(model, conf_control)
opt_control.build()
print(opt_control)

Model pretvorimo v COBRA zapis:

model_control = opt_control.cobra_model()

In mu dodajmo še biomaso:

if require_biomass:
    model_control.add_reactions([model.reactions.BIOMASS_maintenance])
    model_control.objective=model.objective
require_biomass

Model lahko zdaj shranimo v SBML zapis:

cobra.io.write_sbml_model(model_control, os.path.join(f'{folder}','model_control_RECON3D.xml'))

**Branje iz datoteke (če ste prejšnje korake preskočili).**

model_control = cobra.io.read_sbml_model(os.path.join(f'{folder}','model_control_RECON3D.xml'))

model_control.summary()

Shranimo zgornje ukaze kot funkcijo:

def process_opt(model, conf, require_biomass, biomass=None, filename=None, folder=None, reaction = 'BIOMASS_maintenance'):
    react = getattr(model.reactions, reaction)
    react
    
    # integracijo z algoritmom
    if require_biomass:
        opt = CORDA(model, conf, met_prod=biomass)
    else:
        opt = CORDA(model, conf)

    opt.build()
    print(opt)

    # Model pretvorimo v COBRA zapis
    model_cobra = opt.cobra_model()

    # In mu dodajmo še biomaso
    if require_biomass:
        model_cobra.add_reactions([react])
        model_cobra.objective = model.objective

    # Model lahko zdaj shranimo v SBML zapis
    if filename is not None:
        cobra.io.write_sbml_model(model_cobra, os.path.join(f'{folder}', f'{filename}.xml'))
        
    return model_cobra

**POZOR: časovno potratno! Spodnje celice lahko preskočiš in podatke prebereš iz datoteke (spodaj).**

conf_tumor = calculate_confidence(
    df = df_tumor_sub_recon,
    perc0 = perc0,
    perc1 = perc1,
    perc2 = perc2,
    perc3 = perc3,
    model = model
    )

model_tumor = process_opt(
    model = model,
    conf = conf_tumor,
    require_biomass = require_biomass,
    biomass = biomass,
    filename = 'model_tumor_RECON3D',
    folder = folder
)

**Branje iz datoteke (če ste prejšnje korake preskočili).**

model_tumor = cobra.io.read_sbml_model(os.path.join(f'{folder}','model_tumor_RECON3D.xml'))

model_tumor.summary()

### Rangiranje H-GEM

Control:


# id="MAR09931" name="Biomass maintenance reaction without replication precursors"

biomass = model_H_gem.reactions.MAR09931.build_reaction_string().replace("-->", "->")
biomass

conf_control_H_gem = calculate_confidence(
    df = df_control_sub_H_gem,
    perc0 = perc0,
    perc1 = perc1,
    perc2 = perc2,
    perc3 = perc3,
    model = model_H_gem,
    genes = 'ENS_ID'
    )

model_control_H_gem = process_opt(
    model = model_H_gem,
    conf = conf_control_H_gem,
    require_biomass = require_biomass,
    biomass = biomass,
    filename = 'model_control_H-GEM',
    folder = folder,
    reaction = 'MAR09931'
)

model_control_H_gem.summary()

Zaradi dolgega časa izvajanja je debugiranje težavno.

Model summary je poročal:
No objective coefficients in model. Unclear what should be optimized

model_H_gem.reactions.MAR09931.build_reaction_string().replace("-->", "->")

model = model_H_gem
conf = conf_control_H_gem
require_biomass = require_biomass
biomass = biomass
filename = 'model_control_H-GEM'
folder = folder
reaction = 'MAR09931'

react = getattr(model.reactions, reaction)
react



# Model lahko zdaj shranimo v SBML zapis
#if filename is not None:
 #   cobra.io.write_sbml_model(model_cobra, os.path.join(f'{folder}', f'{filename}.xml'))

conf

# integracijo z algoritmom
#if require_biomass:
#    opt = CORDA(model, conf, met_prod=biomass)
#else:
#    opt = CORDA(model, conf)
opt = CORDA(model, conf)

opt.build()
print(opt)

# Model pretvorimo v COBRA zapis
model_cobra = opt.cobra_model()

# In mu dodajmo še biomaso
#if require_biomass:
#    model_cobra.add_reactions([react])
#    model_cobra.objective = model.objective

model_cobra.summary()

model_control_H_gem2 = cobra.io.read_sbml_model(os.path.join(f'{folder}','model_control_H-GEM.xml'))

Tumor:

conf_tumor_H_gem = calculate_confidence(
    df = df_tumor_sub_H_gem,
    perc0 = perc0,
    perc1 = perc1,
    perc2 = perc2,
    perc3 = perc3,
    model = model_H_gem,
    genes = 'ENS_ID'
    )

model_tumor_H_gem = process_opt(
    model = model_H_gem,
    conf = conf_tumor_H_gem,
    require_biomass = require_biomass,
    biomass = biomass,
    filename = 'model_tumor_H-GEM',
    folder = folder,
    reaction = 'MAR09931'
)

model_tumor_H_gem = cobra.io.read_sbml_model(os.path.join(f'{folder}','model_tumor_H-GEM.xml'))

model_tumor_H_gem.summary()
