==> 03_flux_sampling.ipynb <==
import pandas as pd
import cobra

from cobra.io import load_model

import numpy as np
import pickle

import os.path

## Osnovne nastavitve

require_biomass = True
folder = os.path.join('models','biomass') if require_biomass else os.path.join('models','no_biomass')
folder_samples = os.path.join('samples','biomass') if require_biomass else os.path.join('samples','biomass')

## Preberi modele

model_control = cobra.io.read_sbml_model(os.path.join(f'{folder}','model_control_RECON3D.xml'))
model_tumor = cobra.io.read_sbml_model(os.path.join(f'{folder}','model_tumor_RECON3D.xml'))

# H-GEM
# model_control_H_gem = cobra.io.read_sbml_model(os.path.join(f'{folder}','model_control_H-GEM.xml'))
# model_tumor_H_gem = cobra.io.read_sbml_model(os.path.join(f'{folder}','model_tumor_H-GEM.xml'))

Pridobi reakcije v posameznem modelu in jih zapiši kot množica (`set`).

reacts_control = {r.id for r in model_control.reactions}
reacts_tumor = {r.id for r in model_tumor.reactions}

# H-GEM
# reacts_control_H_gem = {r.id for r in model_control_H_gem.reactions}
# reacts_tumor_H_gem = {r.id for r in model_tumor_H_gem.reactions}

## Osnovne analize

### Jaccardov indeks

Jaccardov indeks/koeficient je statistična mera podobnosti dveh populacij/vzorcev.

Izračuna se kot razmerje med številom enakih reakcij v obeh vzorcih (presek množic) in številom vseh različnih reakcij v vzorcih (unija množic):
$$J=\frac{|S_1 \cap S_2|}{|S_1 \cup S_2|}$$

Ker je moč preseka v najboljšem primeru enaka moči unije, se indeks nahaja na intervalu $[0,1]$. Višji Jaccardov indeks pomeni večjo podobnost med opazovanima vzorcema.

J = len(reacts_control & reacts_tumor)/len(reacts_control | reacts_tumor)
print('RECON3D', J)

# H-GEM
# J = len(reacts_control_H_gem & reacts_tumor_H_gem)/len(reacts_control_H_gem | reacts_tumor_H_gem)
# print('H-GEM', J)

## Summary

#### RECON3D

model_control.summary()

model_tumor.summary()

#### H-GEM

model_control_H_gem.summary()

model_tumor_H_gem.summary()

## Vzorčenje metabolnih pretokov

Uporabili bomo funkcijo sample paketa COBRApy ([povezava](https://cobrapy.readthedocs.io/en/latest/sampling.html)).

from cobra.sampling import sample

Podamo model, število vzorcev in metodo vzorčenja. 

**POZOR: Časovno potratno! Podatke lahko prebereš iz datoteke (spodaj).**

s_control = sample(model_control, 1000, method="achr")

# H-GEM
# s_control_H_gem = sample(model_control_H_gem, 1000, method="achr")

**POZOR: Časovno potratno! Podatke lahko prebereš iz datoteke (spodaj).**

s_tumor = sample(model_tumor, 1000, method="achr")

# H-GEM
# s_tumor_H_gem = sample(model_tumor_H_gem, 1000, method="achr")

s_control.to_csv(os.path.join(f'{folder_samples}','samples_control_RECON3D.csv'), index=False)

# H-GEM
# s_control_H_gem.to_csv(os.path.join(f'{folder_samples}','samples_control_H-GEM.csv'), index=False)

s_tumor.to_csv(os.path.join(f'{folder_samples}','samples_tumor_RECON3D.csv'), index=False)

# H-GEM
# s_tumor_H_gem.to_csv(os.path.join(f'{folder_samples}','samples_tumor_H-GEM.csv'), index=False)

**Branje iz datotek (če ste časovne potratne korake izpustili)**

s_control = pd.read_csv(os.path.join(f'{folder_samples}','samples_control_RECON3D.csv'))
s_tumor = pd.read_csv(os.path.join(f'{folder_samples}','samples_tumor_RECON3D.csv'))

# H-GEM
s_control_H_gem = pd.read_csv(os.path.join(f'{folder_samples}','samples_control_H-GEM.csv'))
s_tumor_H_gem = pd.read_csv(os.path.join(f'{folder_samples}','samples_tumor_H-GEM.csv'))

s_control

s_tumor

s_control_H_gem

s_tumor_H_gem
